# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for submitting and tracking EVM token transfers across multiple chains.\nSupports native token and ERC20 transfers with nonce management, gas optimization, and retry logic.",
        "title": "EVM Transaction Sender API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns the operational health of the service, including database connectivity,\nper-chain RPC status and latest block number, and active pipeline queue depths.\nResponds with 200 when all components are healthy, 503 when any component is degraded.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/internal_api.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service is degraded",
                        "schema": {
                            "$ref": "#/definitions/internal_api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "description": "List transactions with optional filtering by chain, sender, or status.\nSupports cursor-based pagination via the `after` query parameter (pass the last transaction ID from the previous page).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by chain ID",
                        "name": "chain_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sender address (hex)",
                        "name": "sender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (QUEUED|PENDING|SUBMITTED|CONFIRMED|REVERTED|FAILED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of results to return (default: 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for pagination: ULID of the last transaction from the previous page",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.TxResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "description": "Fetch a transaction by its ULID identifier, including all broadcast attempts and on-chain receipt data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ULID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "schema": {
                            "$ref": "#/definitions/internal_api.TxResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transfers": {
            "post": {
                "description": "Submit a new native or ERC20 token transfer. Returns immediately with QUEUED status;\nthe transfer is processed asynchronously by the pipeline.\nIdempotency: repeat requests with the same idempotency_key and identical parameters\nreturn the original response without creating a duplicate transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfers"
                ],
                "summary": "Create transfer",
                "parameters": [
                    {
                        "description": "Transfer request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Transfer accepted and queued",
                        "schema": {
                            "$ref": "#/definitions/internal_api.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request (missing fields, bad address, unsupported chain)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Token not whitelisted or disabled",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sender key not registered",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Idempotency key reused with different parameters",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Semantic error (zero amount, amount overflow, exceeds limit)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_apm-dev_evm-tx-sender_internal_pipeline.PipelineInfo": {
            "type": "object",
            "properties": {
                "queue_depth": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_api.AttemptResp": {
            "type": "object",
            "properties": {
                "attempt_number": {
                    "type": "integer"
                },
                "gas_price": {
                    "type": "string"
                },
                "max_fee_per_gas": {
                    "type": "string"
                },
                "max_priority_fee_per_gas": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                }
            }
        },
        "internal_api.ChainHealth": {
            "type": "object",
            "properties": {
                "block_number": {
                    "type": "integer"
                },
                "healthy": {
                    "type": "boolean"
                },
                "rpc_endpoints": {
                    "type": "integer"
                }
            }
        },
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_api.HealthResponse": {
            "type": "object",
            "properties": {
                "chains": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/internal_api.ChainHealth"
                    }
                },
                "db": {
                    "type": "string"
                },
                "pipelines": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_apm-dev_evm-tx-sender_internal_pipeline.PipelineInfo"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_api.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_api.TransferResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "token_contract": {
                    "type": "string"
                },
                "transfer_type": {
                    "type": "string"
                }
            }
        },
        "internal_api.TxResponse": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.AttemptResp"
                    }
                },
                "block_hash": {
                    "type": "string"
                },
                "block_number": {
                    "type": "integer"
                },
                "chain_id": {
                    "type": "integer"
                },
                "confirmations": {
                    "type": "integer"
                },
                "confirmed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "effective_gas_price": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "gas_used": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "nonce": {
                    "type": "integer"
                },
                "receipt_status": {
                    "type": "integer"
                },
                "sender": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "token_contract": {
                    "type": "string"
                },
                "transfer_amount": {
                    "type": "string"
                },
                "transfer_recipient": {
                    "type": "string"
                },
                "transfer_token": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}